Добавлены комментарии к методам
