+++ Yandex.Music.Api/API/YAccountsAPI.cs
@@ -6,10 +6,18 @@ using Yandex.Music.Api.Responses;
 
 namespace Yandex.Music.Api.API
 {
+    /// <summary>
+    /// API для взаимодействия с аккаунтами
+    /// </summary>
     public class YAccountsAPI
     {
         #region Основные функции
 
+        /// <summary>
+        /// Получение списка аккаунтов
+        /// </summary>
+        /// <param name="storage">Хранилище</param>
+        /// <returns></returns>
         public async Task<YAccountResponse> GetAsync(YAuthStorage storage)
         {
             return await new YGetAccountRequest(storage)
@@ -17,6 +25,11 @@ namespace Yandex.Music.Api.API
                 .GetResponseAsync<YAccountResponse>();
         }
 
+        /// <summary>
+        /// Получение списка аккаунтов
+        /// </summary>
+        /// <param name="storage">Хранилище</param>
+        /// <returns></returns>
         public YAccountResponse Get(YAuthStorage storage)
         {
             return GetAsync(storage).GetAwaiter().GetResult();
