+++ Yandex.Music.Api/API/YAlbumAPI.cs
@@ -6,10 +6,19 @@ using Yandex.Music.Api.Responses;
 
 namespace Yandex.Music.Api.API
 {
+    /// <summary>
+    /// API для взаимодействия с альбомами
+    /// </summary>
     public class YAlbumAPI
     {
         #region Основные функции
 
+        /// <summary>
+        /// Получение альбома
+        /// </summary>
+        /// <param name="storage">Хранилище</param>
+        /// <param name="albumId">Идентификатор</param> 
+        /// <returns></returns>
         public async Task<YAlbumResponse> GetAsync(YAuthStorage storage, string albumId)
         {
             return await new YGetAlbumRequest(storage)
@@ -17,6 +26,11 @@ namespace Yandex.Music.Api.API
                 .GetResponseAsync<YAlbumResponse>();
         }
 
+        /// <summary>
+        /// Получение альбома
+        /// </summary>
+        /// <param name="storage">Хранилище</param>
+        /// <param name="albumId">Идентификатор</param>
         public YAlbumResponse Get(YAuthStorage storage, string albumId)
         {
             return GetAsync(storage, albumId).GetAwaiter().GetResult();
