+++ Yandex.Music.Api/API/YLibraryAPI.cs
@@ -6,10 +6,18 @@ using Yandex.Music.Api.Responses;
 
 namespace Yandex.Music.Api.API
 {
+    /// <summary>
+    /// API для взаимодействия с библиотекой
+    /// </summary>
     public class YLibraryAPI
     {
         #region Основные функции
 
+        /// <summary>
+        /// Получение истории
+        /// </summary>
+        /// <param name="storage">Хранилище</param>
+        /// <returns></returns>
         public async Task<YLibraryHistoryResponse> GetHistoryAsync(YAuthStorage storage)
         {
             return await new YGetLibraryHistoryRequest(storage)
@@ -17,11 +25,21 @@ namespace Yandex.Music.Api.API
                 .GetResponseAsync<YLibraryHistoryResponse>();
         }
 
+        /// <summary>
+        /// Получение истории
+        /// </summary>
+        /// <param name="storage">Хранилище</param>
+        /// <returns></returns>
         public YLibraryHistoryResponse GetHistory(YAuthStorage storage)
         {
             return GetHistoryAsync(storage).GetAwaiter().GetResult();
         }
 
+        /// <summary>
+        /// Получение библиотеки
+        /// </summary>
+        /// <param name="storage">Хранилище</param>
+        /// <returns></returns>
         public async Task<YLibraryPlaylistResponse> GetAsync(YAuthStorage storage)
         {
             return await new YGetLibraryPlaylistRequest(storage)
@@ -29,6 +47,11 @@ namespace Yandex.Music.Api.API
                 .GetResponseAsync<YLibraryPlaylistResponse>();
         }
 
+        /// <summary>
+        /// Получение библиотеки
+        /// </summary>
+        /// <param name="storage">Хранилище</param>
+        /// <returns></returns>
         public YLibraryPlaylistResponse Get(YAuthStorage storage)
         {
             return GetAsync(storage).GetAwaiter().GetResult();
