+++ Yandex.Music.Api/API/YPlaylistAPI.cs
@@ -9,6 +9,9 @@ using Yandex.Music.Api.Responses;
 
 namespace Yandex.Music.Api.API
 {
+    /// <summary>
+    /// API для взамодействия с плейлистами
+    /// </summary>
     public class YPlaylistAPI
     {
         #region Вспомогательные функции
@@ -19,6 +22,11 @@ namespace Yandex.Music.Api.API
 
         #region Список с главной
 
+        /// <summary>
+        /// Получение персональных списков
+        /// </summary>
+        /// <param name="storage">Хранилище</param>
+        /// <returns></returns>
         public async Task<List<YPlaylist>> MainPagePersonalAsync(YAuthStorage storage)
         {
             return await new YGetPlaylistMainPageRequest(storage)
@@ -26,6 +34,11 @@ namespace Yandex.Music.Api.API
                 .GetResponseAsyncList<YPlaylist>("$..[?(@.type == 'personal-playlist')].data.data");
         }
 
+        /// <summary>
+        /// Получение персональных списков
+        /// </summary>
+        /// <param name="storage">Хранилище</param>
+        /// <returns></returns>
         public List<YPlaylist> MainPagePersonal(YAuthStorage storage)
         {
             return MainPagePersonalAsync(storage).GetAwaiter().GetResult();
@@ -35,6 +48,11 @@ namespace Yandex.Music.Api.API
 
         #region Стандартные плейлисты
 
+        /// <summary>
+        /// Избранное
+        /// </summary>
+        /// <param name="storage">Хранилище</param>
+        /// <returns></returns>
         public async Task<YPlaylistFavoritesResponse> FavoritesAsync(YAuthStorage storage)
         {
             return await new YGetPlaylistFavoritesRequest(storage)
@@ -42,11 +60,22 @@ namespace Yandex.Music.Api.API
                 .GetResponseAsync<YPlaylistFavoritesResponse>();
         }
 
+        /// <summary>
+        /// Избранное
+        /// </summary>
+        /// <param name="storage">Хранилище</param>
+        /// <returns></returns>
         public YPlaylistFavoritesResponse Favorites(YAuthStorage storage)
         {
             return FavoritesAsync(storage).GetAwaiter().GetResult();
         }
 
+        /// <summary>
+        /// Плейлист дня
+        /// </summary>
+        /// <param name="storage">Хранилище</param>
+        /// <param name="kinds">Тип</param>
+        /// <returns></returns>
         public async Task<YPlaylist> OfTheDayAsync(YAuthStorage storage, string kinds)
         {
             return await new YGetPlaylistOfDayRequest(storage)
@@ -54,11 +83,23 @@ namespace Yandex.Music.Api.API
                 .GetResponseAsync<YPlaylist>("playlist");
         }
 
+        /// <summary>
+        /// Плейлист дня
+        /// </summary>
+        /// <param name="storage">Хранилище</param>
+        /// <param name="kinds">Тип</param>
+        /// <returns></returns>
         public YPlaylist OfTheDay(YAuthStorage storage, string kinds)
         {
             return OfTheDayAsync(storage, kinds).GetAwaiter().GetResult();
         }
 
+        /// <summary>
+        /// Дежавю
+        /// </summary>
+        /// <param name="storage">Хранилище</param>
+        /// <param name="kinds">Тип</param>
+        /// <returns></returns>
         public async Task<YPlaylist> DejaVuAsync(YAuthStorage storage, string kinds)
         {
             return await new YGetPlaylistDejaVuRequest(storage)
@@ -66,11 +107,23 @@ namespace Yandex.Music.Api.API
                 .GetResponseAsync<YPlaylist>("playlist");
         }
 
+        /// <summary>
+        /// Дежавю
+        /// </summary>
+        /// <param name="storage">Хранилище</param>
+        /// <param name="kinds">Тип</param>
+        /// <returns></returns>
         public YPlaylist DejaVu(YAuthStorage storage, string kinds)
         {
             return DejaVuAsync(storage, kinds).GetAwaiter().GetResult();
         }
 
+        /// <summary>
+        /// Премьера
+        /// </summary>
+        /// <param name="storage">Хранилище</param>
+        /// <param name="kinds">Тип</param>
+        /// <returns></returns>
         public async Task<YPlaylist> PremiereAsync(YAuthStorage storage, string kinds)
         {
             return await new YGetPlaylistPremiereRequest(storage)
@@ -78,11 +131,23 @@ namespace Yandex.Music.Api.API
                 .GetResponseAsync<YPlaylist>("playlist");
         }
 
+        /// <summary>
+        /// Премьера
+        /// </summary>
+        /// <param name="storage">Хранилище</param>
+        /// <param name="kinds">Тип</param>
+        /// <returns></returns>
         public YPlaylist Premiere(YAuthStorage storage, string kinds)
         {
             return PremiereAsync(storage, kinds).GetAwaiter().GetResult();
         }
 
+        /// <summary>
+        /// Тайник
+        /// </summary>
+        /// <param name="storage">Хранилище</param>
+        /// <param name="kinds">Тип</param>
+        /// <returns></returns>
         public async Task<YPlaylist> MissedAsync(YAuthStorage storage, string kinds)
         {
             return await new YGetPlaylistMissedRequest(storage)
@@ -90,6 +155,12 @@ namespace Yandex.Music.Api.API
                 .GetResponseAsync<YPlaylist>("playlist");
         }
 
+        /// <summary>
+        /// Тайник
+        /// </summary>
+        /// <param name="storage">Хранилище</param>
+        /// <param name="kinds">Тип</param>
+        /// <returns></returns>
         public YPlaylist Missed(YAuthStorage storage, string kinds)
         {
             return MissedAsync(storage, kinds).GetAwaiter().GetResult();
@@ -99,6 +170,12 @@ namespace Yandex.Music.Api.API
 
         #region Операции над плейлистами
 
+        /// <summary>
+        /// Получение плейлиста
+        /// </summary>
+        /// <param name="storage">Хранилище</param>
+        /// <param name="kinds">Тип</param>
+        /// <returns></returns>
         public async Task<YPlaylist> GetAsync(YAuthStorage storage, string kinds)
         {
             return await new YGetPlaylistRequest(storage)
@@ -106,11 +183,23 @@ namespace Yandex.Music.Api.API
                 .GetResponseAsync<YPlaylist>("playlist");
         }
 
+        /// <summary>
+        /// Получение плейлиста
+        /// </summary>
+        /// <param name="storage">Хранилище</param>
+        /// <param name="kinds">Тип</param>
+        /// <returns></returns>
         public YPlaylist Get(YAuthStorage storage, string kinds)
         {
             return GetAsync(storage, kinds).GetAwaiter().GetResult();
         }
 
+        /// <summary>
+        /// Создание
+        /// </summary>
+        /// <param name="storage">Хранилище</param>
+        /// <param name="name">Заголовок</param>
+        /// <returns></returns>
         public async Task<YPlaylistChangeResponse> CreateAsync(YAuthStorage storage, string name)
         {
             return await new YPlaylistChangeRequest(storage)
@@ -118,11 +207,23 @@ namespace Yandex.Music.Api.API
                 .GetResponseAsync<YPlaylistChangeResponse>();
         }
 
+        /// <summary>
+        /// Создание
+        /// </summary>
+        /// <param name="storage">Хранилище</param>
+        /// <param name="name">Заголовок</param>
+        /// <returns></returns>
         public YPlaylistChangeResponse Create(YAuthStorage storage, string name)
         {
             return CreateAsync(storage, name).GetAwaiter().GetResult();
         }
 
+        /// <summary>
+        /// Удаление
+        /// </summary>
+        /// <param name="storage">Хранилище</param>
+        /// <param name="kind">Тип</param>
+        /// <returns></returns>
         public async Task<bool> RemoveAsync(YAuthStorage storage, string kind)
         {
             try {
@@ -139,11 +240,25 @@ namespace Yandex.Music.Api.API
             return false;
         }
 
+        /// <summary>
+        /// Удаление
+        /// </summary>
+        /// <param name="storage">Хранилище</param>
+        /// <param name="kind">Тип</param>
+        /// <returns></returns>
         public bool Remove(YAuthStorage storage, string kind)
         {
             return RemoveAsync(storage, kind).GetAwaiter().GetResult();
         }
 
+        /// <summary>
+        /// Добавление трека
+        /// </summary>
+        /// <param name="storage">Хранилище</param>
+        /// <param name="trackId">Идентификатор трека</param>
+        /// <param name="albumId">Идентификатор альбома</param>
+        /// <param name="playlistKind">Тип</param>
+        /// <returns></returns>
         public async Task<YInsertTrackToPlaylistResponse> InsertTrackAsync(YAuthStorage storage, string trackId, string albumId,
             string playlistKind)
         {
@@ -152,11 +267,28 @@ namespace Yandex.Music.Api.API
                 .GetResponseAsync<YInsertTrackToPlaylistResponse>();
         }
 
+        /// <summary>
+        /// Добавление трека
+        /// </summary>
+        /// <param name="storage">Хранилище</param>
+        /// <param name="trackId">Идентификатор трека</param>
+        /// <param name="albumId">Идентификатор альбома</param>
+        /// <param name="playlistKind">Тип</param>
+        /// <returns></returns>
         public YInsertTrackToPlaylistResponse InsertTrack(YAuthStorage storage, string trackId, string albumId, string playlistKind)
         {
             return InsertTrackAsync(storage, trackId, albumId, playlistKind).GetAwaiter().GetResult();
         }
 
+        /// <summary>
+        /// Удаление треков
+        /// </summary>
+        /// <param name="storage">Хранилище</param>
+        /// <param name="from">Начало интервала</param>
+        /// <param name="to">Конец интервала</param>
+        /// <param name="revision">Ревизия</param>
+        /// <param name="playlistKind">Тип</param>
+        /// <returns></returns>
         public async Task<YDeleteTrackFromPlaylistResponse> DeleteTrackAsync(YAuthStorage storage, int from, int to, int revision,
             string playlistKind)
         {
@@ -165,6 +297,15 @@ namespace Yandex.Music.Api.API
                 .GetResponseAsync<YDeleteTrackFromPlaylistResponse>();
         }
 
+        /// <summary>
+        /// Удаление треков
+        /// </summary>
+        /// <param name="storage">Хранилище</param>
+        /// <param name="from">Начало интервала</param>
+        /// <param name="to">Конец интервала</param>
+        /// <param name="revision">Ревизия</param>
+        /// <param name="playlistKind">Тип</param>
+        /// <returns></returns>
         public YDeleteTrackFromPlaylistResponse DeleteTrack(YAuthStorage storage, int from, int to, int revision, string playlistKind)
         {
             return DeleteTrackAsync(storage, from, to, revision, playlistKind).GetAwaiter().GetResult();
