+++ Yandex.Music.Api/API/YUserAPI.cs
@@ -4,6 +4,7 @@ using System.Threading.Tasks;
 using Yandex.Music.Api.Common;
 using Yandex.Music.Api.Requests;
 using Yandex.Music.Api.Requests.Auth;
+using Yandex.Music.Api.Requests.Yandex;
 using Yandex.Music.Api.Responses;
 
 namespace Yandex.Music.Api.API
@@ -109,6 +110,18 @@ namespace Yandex.Music.Api.API
         public YAuthInfoUserResponse GetUserAuthDetails(YAuthStorage storage)
         {
             return GetUserAuthDetailsAsync(storage).GetAwaiter().GetResult();
+        }
+
+        public async Task<YGetCookieResponse> GetYandexCookieAsync(YAuthStorage storage)
+        {
+            return await new YGetCookieRequest(storage)
+                .Create()
+                .GetResponseAsync<YGetCookieResponse>();
+        }
+
+        public YGetCookieResponse GetYandexCookie(YAuthStorage storage)
+        {
+            return GetYandexCookieAsync(storage).GetAwaiter().GetResult();
         }
 
         #endregion Основные функции
