+++ Yandex.Music.Api/Common/YSearchType.cs
@@ -1,0 +1,29 @@
+﻿namespace Yandex.Music.Api.Common
+{
+    /// <summary>
+    /// Тип объекта поиска
+    /// </summary>
+    public enum YSearchType
+    {
+        /// <summary>
+        /// Артисты
+        /// </summary>
+        Artists,
+        /// <summary>
+        /// Альбомы
+        /// </summary>
+        Albums,
+        /// <summary>
+        /// Плейлисты
+        /// </summary>
+        Playlists,
+        /// <summary>
+        /// Треки
+        /// </summary>
+        Tracks,
+        /// <summary>
+        /// Пользователи
+        /// </summary>
+        Users
+    }
+}
