--- Yandex.Music.Api/Common/YandexSearchType.cs
@@ -1,11 +1,0 @@
-﻿namespace Yandex.Music.Api.Common
-{
-    public enum YandexSearchType
-    {
-        Tracks,
-        Artists,
-        Albums,
-        Playlists,
-        Users
-    }
-}
