+++ Yandex.Music.Api/Responses/YGetCookieResponse.cs
@@ -2,7 +2,7 @@ namespace Yandex.Music.Api.Responses
 {
     public class YGetCookieResponse
     {
-        public string CryptoId { get; set; }
-        public string CryptoSign { get; set; }
+        public string CryptoUId { get; set; }
+        public string CryptoUId_Sign { get; set; }
     }
 }
